Move display/enhancements.py to _internal/
