+++ certbot/_internal/client.py
@@ -30,7 +30,7 @@ from certbot import interfaces
 from certbot._internal import storage
 from certbot import util
 from certbot.compat import os
-from certbot.display import enhancements
+from certbot._internal.display import enhancements
 from certbot.display import ops as display_ops
 from certbot.plugins import selection as plugin_selection
